Factorise l'affichage des situations.
